requirements.txt
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
Pillow>=10.0.0
